Second Commit after watching end to end service video

+++ AccountSystemNew/AccountSystemNewRepository/src/main/java/za/ac/nwu/acnew/repo/persistence/AccountTypeRepository.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Repository;
 import za.ac.nwu.acnew.domain.persistence.AccountType;
 
 @Repository
-public interface AccountTypeRepository extends JpaRepository<AccountType, Integer> {
+public interface AccountTypeRepository extends JpaRepository<AccountType, Long> {
+
 }
